+++ lib/nimble_template/template.ex
@@ -1,24 +1,20 @@
 defmodule NimbleTemplate.Template do
   @moduledoc false
 
-  alias NimbleTemplate.Api.Template, as: ApiTemplate
-  alias NimbleTemplate.Live.Template, as: LiveTemplate
   alias NimbleTemplate.Mix.Template, as: MixTemplate
-  alias NimbleTemplate.Web.Template, as: WebTemplate
+  alias NimbleTemplate.Phoenix.Template, as: PhoenixTemplate
   alias NimbleTemplate.{Addons, Project}
 
   def apply(%Project{mix_project?: true} = project) do
     MixTemplate.apply(project)
 
-    if Mix.shell().yes?("\nFetch and install dependencies?"), do: Mix.shell().cmd("mix deps.get")
+    fetch_and_install_dependencies()
   end
 
   def apply(%Project{} = project) do
-    project
-    |> common_setup()
-    |> variant_setup()
+    PhoenixTemplate.apply(project)
 
-    if Mix.shell().yes?("\nFetch and install dependencies?"), do: Mix.shell().cmd("mix deps.get")
+    fetch_and_install_dependencies()
   end
 
   def host_on_github?(), do: Mix.shell().yes?("\nWill you host this project on Github?")
@@ -35,40 +31,7 @@ defmodule NimbleTemplate.Template do
   def install_addon_prompt?(addon),
     do: Mix.shell().yes?("\nWould you like to add the #{addon} addon?")
 
-  # Common setup for both API and Web projects
-  defp common_setup(%Project{} = project) do
-    project
-    |> Addons.ElixirVersion.apply()
-    |> Addons.Readme.apply()
-    |> Addons.Makefile.apply()
-    |> Addons.Docker.apply()
-    |> Addons.MixRelease.apply()
-    |> Addons.TestEnv.apply()
-    |> Addons.Credo.apply()
-    |> Addons.Dialyxir.apply()
-    |> Addons.ExCoveralls.apply()
-    |> Addons.ExMachina.apply()
-    |> Addons.Mimic.apply()
-
-    if host_on_github?() do
-      if generate_github_template?(),
-        do: Addons.Github.apply(project, %{github_template: true})
-
-      if generate_github_action?(),
-        do: Addons.Github.apply(project, %{github_action: true})
-    end
-
-    if install_addon_prompt?("Oban"), do: Addons.Oban.apply(project)
-    if install_addon_prompt?("ExVCR"), do: Addons.ExVCR.apply(project)
-
-    project
+  defp fetch_and_install_dependencies() do
+    if Mix.shell().yes?("\nFetch and install dependencies?"), do: Mix.shell().cmd("mix deps.get")
   end
-
-  defp variant_setup(%Project{api_project?: true} = project), do: ApiTemplate.apply(project)
-
-  defp variant_setup(%Project{web_project?: true, live_project?: false} = project),
-    do: WebTemplate.apply(project)
-
-  defp variant_setup(%Project{web_project?: true, live_project?: true} = project),
-    do: LiveTemplate.apply(project)
 end
